Separate endpoint for OAuth and removed get_note_by_id
